+++ machines/hostup1/hardware.nix
@@ -40,4 +40,7 @@
   # Guest additions for Qemu.
   services.spice-vdagentd.enable = true;
 
+  # Allow building ARM 64-bit via QEMU.
+  boot.binfmt.emulatedSystems = [ "aarch64-linux" ];
+
 }